pCMORizer.f90 fix: no NaN flds + corrupted netCDF
When using raw model output with timesteps from 00-23 per file, a
subsequent file is needed to allow for the calculation of cell_method
mean and sum fields, such as rsds, hfls, pr, etc. Therefore subsequent
files are always part of the filelist the CMORizer works on. At the same
time the CMORizer checks the time information in raw model outputs and
creates -- depending on configuration -- the output files. In case an
output file exists, such as of the year+1, the data is
processed and written to the year+1; i.e., this leads to fields being
processed at the end of the processing for year and written to the shared
CMORized netCDF files of year+1. In case of point data this does not do
any harm. In case of mean data fields, there is always the subsequent
time step missing at the very end of the last day of the files in the
filelist of year. This leads to missing data at -01-01T23:30:00Z in
case of daily raw inputs. At the same the the write operation from the
processing of year into files produced by the processing for year+1 may
laead to corrupted netCDF files, with HDF5 errors. (If raw data is
stored with the subsequent timestep of the consecutive year, this is not
happening, 00 until d+1_00.) To prevent this behaviour from happening,
an additional check has been added and tested, which checks whether a
variable at a given timestep has been  written / exists already. In
this case the field is not written once more. This additional check
causes minimum overhead to the processing. This is especially an issue
with multiple years being processed at the same time.