Remove last file from tge loop ove list of Files
The code loop over all the files in the directory. This is
a quick fix to avoid recreating a new yearly file, which will
be filled with the missing value for the accumulated variables
if the code is run for the years in reverse order, e.g. first
run 2013, then for 2014. To avoid this, it is necessary to remove
reading the last file, since the last file is from the next year
(1. 1. next year), and the code created a file for the next year
or refills de file with new values, which will inlcude a timestep
witl missing values.