Replace with aggregate_per_day.sh
This script aggreagtes per 3hr, 6hr, and day. The main fortran code
works well for 3hr and 6hr output, but not for daily output. This
script has been replaced with aggregate_per_day.sh, which is more
general, and works for all variables and aggregation methods given in
a csv file.